+++ packages/face-detection/src/web.ts
@@ -10,13 +10,13 @@ export class FaceDetectionWeb extends WebPlugin implements FaceDetectionPlugin {
   public async processImage(
     _options: ProcessImageOptions,
   ): Promise<ProcessImageResult> {
-    throw this.createUnavailableException();
+    throw this.createUnimplementedException();
   }
 
-  private createUnavailableException(): CapacitorException {
+  private createUnimplementedException(): CapacitorException {
     return new CapacitorException(
-      'This Face Detection plugin method is not available on this platform.',
-      ExceptionCode.Unavailable,
+      'This method is not implemented on web.',
+      ExceptionCode.Unimplemented,
     );
   }
 }