+++ packages/document-scanner/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 8.0.1
+
+### Patch Changes
+
+- [`98affd15ab9879702ff277e4487f40af14253e91`](https://github.com/capawesome-team/capacitor-mlkit/commit/98affd15ab9879702ff277e4487f40af14253e91) ([#321](https://github.com/capawesome-team/capacitor-mlkit/pull/321)): fix(android): AGP 9.0 no longer supports `proguard-android.txt`
+
+## 8.0.0
+
+### Major Changes
+
+- [`36dc42aeee4af509e84d2904d4be1a6a87f78fac`](https://github.com/capawesome-team/capacitor-mlkit/commit/36dc42aeee4af509e84d2904d4be1a6a87f78fac) ([#313](https://github.com/capawesome-team/capacitor-mlkit/pull/313)): feat!: update dependencies to the latest versions (see `BREAKING.md`)
+
+- [`7c34e5f74f0339aa0b1a210f0060ccb31d0720bd`](https://github.com/capawesome-team/capacitor-mlkit/commit/7c34e5f74f0339aa0b1a210f0060ccb31d0720bd): fix!: throw correct error code (see `BREAKING.md`)
+
+- [`e5a827f63dd47ae190a189760a484f7f5403fca3`](https://github.com/capawesome-team/capacitor-mlkit/commit/e5a827f63dd47ae190a189760a484f7f5403fca3) ([#311](https://github.com/capawesome-team/capacitor-mlkit/pull/311)): feat!: update to Capacitor 8 (see `BREAKING.md`)
+
+## 7.5.0
+
+## 7.4.0
+
 ## 7.3.0
 
 ### Minor Changes